lsp: Add `make_completion_context` testing helper
